Refactor App component to utilize custom hooks for improved state management
- Replaced direct state and effect management with a custom `useApp` hook for better organization and readability.
- Removed legacy migration logic and conversation handling from the App component, streamlining the codebase.
- Updated import paths to include the new `useApp` hook for enhanced modularity.
