+++ src/hooks/index.ts
@@ -10,3 +10,4 @@ export * from "./useHistory";
 export * from "./useCopyToClipboard";
 export * from "./useTitles";
 export * from "./useSystemPrompts";
+export * from "./useApp";
